+++ mowl/evaluation/__init__.py
@@ -1,3 +1,4 @@
-from mowl.evaluation.base import AxiomsRankBasedEvaluator
+from mowl.evaluation.base import Evaluator, AxiomsRankBasedEvaluator
+from mowl.evaluation.rank_based import RankBasedEvaluator, ModelRankBasedEvaluator, EmbeddingsRankBasedEvaluator
 
-__all__ = ["AxiomsRankBasedEvaluator"]
+